+++ application/src/main/java/org/opentripplanner/model/plan/walkstep/verticaltransportationuse/StairsUse.java
@@ -9,10 +9,10 @@ import javax.annotation.Nullable;
 public class StairsUse extends InclinedVerticalTransportationUse {
 
   public StairsUse(
-    @Nullable Float fromLevel,
+    @Nullable Double fromLevel,
     @Nullable String fromLevelName,
     InclineType inclineType,
-    @Nullable Float toLevel,
+    @Nullable Double toLevel,
     @Nullable String toLevelName
   ) {
     super(fromLevel, fromLevelName, inclineType, toLevel, toLevelName);
